+++ packages/mobile/src/navigation/__figma__/TopNavBar.figma.tsx
@@ -4,7 +4,6 @@ import { Divider, HStack, VStack } from '../../layout';
 import {
   BrowserBar,
   BrowserBarSearchInput,
-  NavBarIconButton,
   NavigationSubtitle,
   NavigationTitle,
   NavigationTitleSelect,
@@ -19,12 +18,12 @@ figma.connect(
       type: 'title + subtitle',
     },
     imports: [
-      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@coinbase/cds-mobile/layout/Divider';",
-      "import { VStack } from '@coinbase/cds-mobile/layout/VStack';",
-      "import { HStack } from '@coinbase/cds-mobile/layout/HStack';",
-      "import { NavigationTitle } from '@coinbase/cds-mobile/navigation/NavigationTitle';",
-      "import { NavigationSubtitle } from '@coinbase/cds-mobile/navigation/NavigationSubtitle';",
+      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@coinbase/cds-mobile/layout/Divider'",
+      "import { VStack } from '@coinbase/cds-mobile/layout/VStack'",
+      "import { HStack } from '@coinbase/cds-mobile/layout/HStack'",
+      "import { NavigationTitle } from '@coinbase/cds-mobile/navigation/NavigationTitle'",
+      "import { NavigationSubtitle } from '@coinbase/cds-mobile/navigation/NavigationSubtitle'",
     ],
     props: {
       title: figma.string('↳ title'),
@@ -34,19 +33,19 @@ figma.connect(
       }),
 
       startAction: figma.boolean('show left action', {
-        true: <NavBarIconButton name="appSwitcher" />,
+        true: figma.instance('left action'),
         false: undefined,
       }),
       endAction1: figma.boolean('show right action', {
-        true: <NavBarIconButton name="more" />,
+        true: figma.instance('right action'),
         false: undefined,
       }),
       endAction2: figma.boolean('show 2nd right action', {
-        true: <NavBarIconButton name="gear" />,
+        true: figma.instance('2nd right action'),
         false: undefined,
       }),
       endAction3: figma.boolean('show 3rd right action', {
-        true: <NavBarIconButton name="share" />,
+        true: figma.instance('3rd right action'),
         false: undefined,
       }),
       divider: figma.boolean('show divider', {
@@ -95,28 +94,28 @@ figma.connect(
       type: 'dropdown',
     },
     imports: [
-      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@coinbase/cds-mobile/layout/Divider';",
-      "import { VStack } from '@coinbase/cds-mobile/layout/VStack';",
-      "import { HStack } from '@coinbase/cds-mobile/layout/HStack';",
-      "import { NavigationTitleSelect } from '@coinbase/cds-mobile/navigation/NavigationTitleSelect';",
+      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@coinbase/cds-mobile/layout/Divider'",
+      "import { VStack } from '@coinbase/cds-mobile/layout/VStack'",
+      "import { HStack } from '@coinbase/cds-mobile/layout/HStack'",
+      "import { NavigationTitleSelect } from '@coinbase/cds-mobile/navigation/NavigationTitleSelect'",
     ],
     props: {
       title: figma.string('↳ title'),
       startAction: figma.boolean('show left action', {
-        true: <NavBarIconButton name="appSwitcher" />,
+        true: figma.instance('left action'),
         false: undefined,
       }),
       endAction1: figma.boolean('show right action', {
-        true: <NavBarIconButton name="more" />,
+        true: figma.instance('right action'),
         false: undefined,
       }),
       endAction2: figma.boolean('show 2nd right action', {
-        true: <NavBarIconButton name="gear" />,
+        true: figma.instance('2nd right action'),
         false: undefined,
       }),
       endAction3: figma.boolean('show 3rd right action', {
-        true: <NavBarIconButton name="share" />,
+        true: figma.instance('3rd right action'),
         false: undefined,
       }),
       divider: figma.boolean('show divider', {
@@ -158,27 +157,27 @@ figma.connect(
       type: 'with search',
     },
     imports: [
-      "import { Divider } from '@coinbase/cds-mobile/layout/Divider';",
-      "import { VStack } from '@coinbase/cds-mobile/layout/VStack';",
-      "import { HStack } from '@coinbase/cds-mobile/layout/HStack';",
-      "import { BrowserBar } from '@coinbase/cds-mobile/navigation/BrowserBar';",
-      "import { BrowserBarSearchInput } from '@coinbase/cds-mobile/navigation/BrowserBarSearchInput';",
+      "import { Divider } from '@coinbase/cds-mobile/layout/Divider'",
+      "import { VStack } from '@coinbase/cds-mobile/layout/VStack'",
+      "import { HStack } from '@coinbase/cds-mobile/layout/HStack'",
+      "import { BrowserBar } from '@coinbase/cds-mobile/navigation/BrowserBar'",
+      "import { BrowserBarSearchInput } from '@coinbase/cds-mobile/navigation/BrowserBarSearchInput'",
     ],
     props: {
       startAction: figma.boolean('show left action', {
-        true: <NavBarIconButton name="appSwitcher" />,
+        true: figma.instance('left action'),
         false: undefined,
       }),
       endAction1: figma.boolean('show right action', {
-        true: <NavBarIconButton name="more" />,
+        true: figma.instance('right action'),
         false: undefined,
       }),
       endAction2: figma.boolean('show 2nd right action', {
-        true: <NavBarIconButton name="gear" />,
+        true: figma.instance('2nd right action'),
         false: undefined,
       }),
       endAction3: figma.boolean('show 3rd right action', {
-        true: <NavBarIconButton name="share" />,
+        true: figma.instance('3rd right action'),
         false: undefined,
       }),
       divider: figma.boolean('show divider', {
@@ -216,26 +215,81 @@ figma.connect(
       type: 'empty',
     },
     imports: [
-      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@coinbase/cds-mobile/layout/Divider';",
-      "import { VStack } from '@coinbase/cds-mobile/layout/VStack';",
-      "import { HStack } from '@coinbase/cds-mobile/layout/HStack';",
+      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@coinbase/cds-mobile/layout/Divider'",
+      "import { VStack } from '@coinbase/cds-mobile/layout/VStack'",
+      "import { HStack } from '@coinbase/cds-mobile/layout/HStack'",
     ],
     props: {
       startAction: figma.boolean('show left action', {
-        true: <NavBarIconButton name="appSwitcher" />,
+        true: figma.instance('left action'),
         false: undefined,
       }),
       endAction1: figma.boolean('show right action', {
-        true: <NavBarIconButton name="more" />,
+        true: figma.instance('right action'),
         false: undefined,
       }),
       endAction2: figma.boolean('show 2nd right action', {
-        true: <NavBarIconButton name="gear" />,
+        true: figma.instance('2nd right action'),
         false: undefined,
       }),
       endAction3: figma.boolean('show 3rd right action', {
-        true: <NavBarIconButton name="share" />,
+        true: figma.instance('3rd right action'),
+        false: undefined,
+      }),
+      divider: figma.boolean('show divider', {
+        true: <Divider />,
+        false: undefined,
+      }),
+    },
+    example: function Example({ startAction, endAction1, endAction2, endAction3, divider }) {
+      return (
+        <VStack>
+          <TopNavBar
+            end={
+              <HStack alignItems="center" gap={0.25}>
+                {endAction3}
+                {endAction2}
+                {endAction1}
+              </HStack>
+            }
+            start={startAction}
+          />
+          {divider}
+        </VStack>
+      );
+    },
+  },
+);
+
+figma.connect(
+  TopNavBar,
+  'https://www.figma.com/design/k5CtyJccNQUGMI5bI4lJ2g/%E2%9C%A8-CDS-Components?node-id=49598-4137',
+  {
+    variant: {
+      type: 'stepper',
+    },
+    imports: [
+      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@coinbase/cds-mobile/layout/Divider'",
+      "import { VStack } from '@coinbase/cds-mobile/layout/VStack'",
+      "import { HStack } from '@coinbase/cds-mobile/layout/HStack'",
+    ],
+    props: {
+      startAction: figma.boolean('show left action', {
+        true: figma.instance('left action'),
+        false: undefined,
+      }),
+      endAction1: figma.boolean('show right action', {
+        true: figma.instance('right action'),
+        false: undefined,
+      }),
+      endAction2: figma.boolean('show 2nd right action', {
+        true: figma.instance('2nd right action'),
+        false: undefined,
+      }),
+      endAction3: figma.boolean('show 3rd right action', {
+        true: figma.instance('3rd right action'),
         false: undefined,
       }),
       divider: figma.boolean('show divider', {
@@ -271,27 +325,27 @@ figma.connect(
       type: 'Market Selector',
     },
     imports: [
-      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar';",
-      "import { Divider } from '@coinbase/cds-mobile/layout/Divider';",
-      "import { VStack } from '@coinbase/cds-mobile/layout/VStack';",
-      "import { HStack } from '@coinbase/cds-mobile/layout/HStack';",
+      "import { TopNavBar } from '@coinbase/cds-mobile/navigation/TopNavBar'",
+      "import { Divider } from '@coinbase/cds-mobile/layout/Divider'",
+      "import { VStack } from '@coinbase/cds-mobile/layout/VStack'",
+      "import { HStack } from '@coinbase/cds-mobile/layout/HStack'",
     ],
     props: {
-      children: figma.children('Market selector'),
+      children: figma.children('SelectChip'),
       startAction: figma.boolean('show left action', {
-        true: <NavBarIconButton name="appSwitcher" />,
+        true: figma.instance('left action'),
         false: undefined,
       }),
       endAction1: figma.boolean('show right action', {
-        true: <NavBarIconButton name="more" />,
+        true: figma.instance('right action'),
         false: undefined,
       }),
       endAction2: figma.boolean('show 2nd right action', {
-        true: <NavBarIconButton name="gear" />,
+        true: figma.instance('2nd right action'),
         false: undefined,
       }),
       endAction3: figma.boolean('show 3rd right action', {
-        true: <NavBarIconButton name="share" />,
+        true: figma.instance('3rd right action'),
         false: undefined,
       }),
       divider: figma.boolean('show divider', {