+++ packages/web-visualization/CHANGELOG.md
@@ -8,13 +8,16 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
-## 3.4.0-beta.13 (1/13/2026 PST)
+## 3.4.0-beta.13 (1/20/2026 PST)
 
-#### 🐞 Fixes
+#### 🚀 Updates
 
 - Add chart Legend component. [[#302](https://github.com/coinbase/cds/pull/302)]
 - Add support for hideBeaconLabels in Scrubber. [[#302](https://github.com/coinbase/cds/pull/302)]
-- Add support for custom bar components. [[#302](https://github.com/coinbase/cds/pull/302)]
+
+#### 📘 Misc
+
+- Internal: code connect file lint fixes. [[#311](https://github.com/coinbase/cds/pull/311)]
 
 ## 3.4.0-beta.12 (1/8/2026 PST)